requirements.txt
tensorflow-cpu==2.13.0
Pillow==10.0.0
numpy==1.24.3
ImageHash==4.3.1
